refactor: share benchmark metric post-processing
